# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyTravel API",
        "description": "Documentation for EasyTravel project.",
        "contact": {
            "email": "21732184@live.uem.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/countries/countries": {
            "get": {
                "tags": [
                    "countries"
                ],
                "summary": "Get all countries",
                "description": "Returns all the countries of the database.",
                "operationId": "getCountries",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All models",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": 1,
                                            "name": "Afghanistan"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Albania"
                                        },
                                        {
                                            "id": 3,
                                            "name": "Algeria"
                                        },
                                        {
                                            "id": 4,
                                            "name": "American Samoa"
                                        },
                                        {
                                            "id": 5,
                                            "name": "Angola"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}